requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
alembic
passlib
python-jose[cryptography]
python-dotenv
bcrypt
